alg_code/cli_interface.py
```python
import sys
import os
import re
import json
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

def load_shakespeare(filepath):
    try:
        with open(filepath, 'r') as file:
            shakespeare_works = file.read()
            return shakespeare_works
    except FileNotFoundError:
        print('The file was not found.')
        sys.exit(1)

def get_user_input():
    user_input = input('Please enter your text: ')
    return user_input

def preprocess_text(text):
    text = re.sub(r'[^A-Za-z ]+', '', text)
    text = text.lower()
    return text

def compare_to_shakespeare(user_input, shakespeare_works):
    user_input_processed = preprocess_text(user_input)
    shakespeare_processed = preprocess_text(shakespeare_works)
    corpus = [shakespeare_processed, user_input_processed]
    vectorizer = TfidfVectorizer()
    tfidf_matrix = vectorizer.fit_transform(corpus)
    cosine_sim = cosine_similarity(tfidf_matrix[0:1], tfidf_matrix[1:2])
    similarity_score = cosine_sim[0][0]
    return similarity_score

if __name__ == '__main__':
    try:
        works = load_shakespeare('shakespeare.txt')
        user_input = get_user_input()
        if user_input.strip() == '':
            raise ValueError('Input is empty.')
        if len(user_input) > 1000:
            raise ValueError('Input is too long.')
        if not user_input.isprintable():
            raise ValueError('Input contains invalid characters.')
        score = compare_to_shakespeare(user_input, works)
        result_json = json.dumps({'similarityScore': score})
        print(result_json)
    except ValueError as ve:
        error_json = json.dumps({'error': str(ve)})
        print(error_json)
    except KeyboardInterrupt:
        print('\nProcess interrupted by user.')
        sys.exit(0)
    except Exception as e:
        error_json = json.dumps({'error': 'An unexpected error occurred.'})
        print(error_json)
        sys.exit(1)
```