alg_code/test_cli_interface.py
```python
import unittest
from unittest.mock import patch

# Assuming the existence of these functions in cli_interface.py
# def get_user_input():
#     ...
# def load_shakespeare(filename):
#     ...
# def preprocess_text(text):
#     ...
# def compare_to_shakespeare(user_input, shakespeare_text):
#     ...

class TestCLIInterface(unittest.TestCase):

    @patch('builtins.input', return_value='Hello World')
    def test_get_user_input(self, mock_input):
        user_input = get_user_input()
        self.assertEqual(user_input, 'Hello World')

    def test_load_shakespeare_file_not_found(self):
        with self.assertRaises(SystemExit):
            load_shakespeare('non_existent_file.txt')

    def test_preprocess_text(self):
        result = preprocess_text('Hello! World,')
        expected = 'hello world'
        self.assertEqual(result, expected)

    def test_compare_to_shakespeare(self):
        dummy_shakespeare = 'To be, or not to be, that is the question:'
        dummy_input = 'To be or not to be'
        score = compare_to_shakespeare(dummy_input, dummy_shakespeare)
        self.assertGreaterEqual(score, 0)
        self.assertLessEqual(score, 1)

if __name__ == '__main__':
    unittest.main()
```